게시판 목록 가져오는 함수도 utility 함수로 빼기
